Refactor catalog package with interface
